EPMLSTR-435 Added tests

+++ src/main/java/com/epam/test_generator/services/PasswordService.java
@@ -4,6 +4,7 @@ import com.epam.test_generator.dao.interfaces.TokenDAO;
 import com.epam.test_generator.dto.PasswordResetDTO;
 import com.epam.test_generator.entities.Token;
 import com.epam.test_generator.entities.User;
+import com.epam.test_generator.services.exceptions.TokenMissingException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -24,13 +25,14 @@ public class PasswordService {
 
     public String createResetUrl(HttpServletRequest request, Token token) {
         String url =
-            request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort();
+                request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort();
         String resetUrl = url + "/cucumber/passwordReset?token=" + token.getToken();
         return resetUrl;
     }
+
     public String createConfirmUrl(HttpServletRequest request, Token token) {
         String url =
-            request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort();
+                request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort();
         String resetUrl = url + "/cucumber/confirmAccount?token=" + token.getToken();
         return resetUrl;
     }
@@ -39,13 +41,16 @@ public class PasswordService {
     public void passwordReset(PasswordResetDTO passwordResetDTO) {
         String token = passwordResetDTO.getToken();
         Token resetToken = tokenDAO.findByToken(token);
+        if (resetToken == null) {
+            throw new TokenMissingException("Token is invalid");
+        }
         User user = resetToken.getUser();
         String updatedPassword = passwordEncoder.encode(passwordResetDTO.getPassword());
         userService.updatePassword(updatedPassword, user.getEmail());
         tokenDAO.delete(resetToken);
     }
 
-    public Token getTokenByName(String token){
+    public Token getTokenByName(String token) {
         return tokenDAO.findByToken(token);
     }
 }
